튜토리얼 카드덱 조회 API 추가

+++ src/main/java/ssho/api/core/domain/userswipe/model/UserSwipeScore.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class UserSwipeScore {
     private String userId;
-    private int[] scoreList;
+    private double[] scoreList;
 }
